fix: resolve issue 284
