+++ src/SlimFaas/Kubernetes/Models/ReplicaRequest.cs
@@ -1,0 +1,3 @@
+namespace SlimFaas.Kubernetes;
+
+public record ReplicaRequest(string Deployment, string Namespace, int Replicas, PodType PodType);
